Reorganized code, fixed problems, added new string methods
