+++ changes.tmp
@@ -49,3 +49,6 @@ Applied	NumeRe supports now UTF8 encoding in code, which means, you can now use 
 Cleaned	The terminal now supports non-ASCII characters as well. Additionally, the selection appearance was unified with the editor appearance (only grey background).
 Added	Plots can now support CJK ideographs, if either "noto" or "notosf" is used as plot font. Loading this font can take some seconds due to its file size, so please be patient.
 Added	"stix" and "stix2" has been added as plotting font.
+Cleaned	String functions do now interact on bytes, which are only identical to characters in the ASCII-only case. Notable exception is the usage of characters a function inputs (e.g. "strmatch()"). Those will be handled as characters instead.
+Added	Methods for handling characters and unicode have been added to the string object: "STR.chr()", "STR.unicodelen", "STR.normalize()", "STR.collate()".
+Applied	The function "char()" has been renamed to "byte()", because that provides a better indication of its functionality. The old name is still interpreted as alias, but shall not be used anymore.
