+++ kernel/core/io/file.hpp
@@ -1620,7 +1620,7 @@ namespace NumeRe
             int32_t versionMinor;
             int32_t versionBuild;
             const short fileSpecVersionMajor = 4;
-            const short fileSpecVersionMinor = 1;
+            const short fileSpecVersionMinor = 10;
             float fileVersionRead;
             size_t checkPos;
             size_t checkStart;
