+++ kernel/core/strings/functionimplementation.cpp
@@ -26,6 +26,7 @@
 #include "../../../database/dbinternals.hpp"
 #include "../../kernel.hpp"
 #include "../../versioninformation.hpp"
+#include <boost/locale.hpp>
 #endif
 
 #include <regex>
@@ -178,6 +179,58 @@ mu::Array strfnc_utf8ToAnsi(const mu::Array& a)
 mu::Array strfnc_ansiToUtf8(const mu::Array& a)
 {
     return mu::apply(ansiToUtf8, a);
+}
+
+
+/////////////////////////////////////////////////
+/// \brief Implementation of the to_codepoints()
+/// function.
+///
+/// \param a const mu::Array&
+/// \return mu::Array
+///
+/////////////////////////////////////////////////
+mu::Array strfnc_to_codepoints(const mu::Array& a)
+{
+    mu::Array ret;
+    ret.copyDims(a);
+
+    for (size_t i = 0; i < a.size(); i++)
+    {
+        mu::Array current;
+        std::u32string u32str = boost::locale::conv::utf_to_utf<char32_t>(a.get(i).getStr());
+        current.reserve(u32str.length());
+
+        for (size_t j = 0; j < u32str.length(); j++)
+        {
+            current.emplace_back((uint32_t)u32str[j]);
+        }
+
+        ret.emplace_back(current);
+    }
+
+    return ret;
+}
+
+
+/////////////////////////////////////////////////
+/// \brief Implementation of the from_codepoints()
+/// function.
+///
+/// \param a const mu::Array&
+/// \return mu::Array
+///
+/////////////////////////////////////////////////
+mu::Array strfnc_from_codepoints(const mu::Array& a)
+{
+    std::u32string u32str;
+
+    for (size_t i = 0; i < a.size(); i++)
+    {
+        u32str.push_back((char32_t)a.get(i).getNum().asUI64());
+    }
+
+    return mu::Value(boost::locale::conv::utf_to_utf<char>(u32str));
 }
 
 
@@ -1542,9 +1595,11 @@ mu::Array strfnc_strmatchall(const mu::Array& chars, const mu::Array& where, con
         else
             pos_last = whereView.get(i).getStr().length()-1;
 
-        for (size_t j = 0; j < charsView.get(i).getStr().length(); j++)
+        std::vector<std::string> vUtf8Chars = splitUtf8Chars(charsView.get(i).getStr());
+
+        for (const std::string& utf8Char : vUtf8Chars)
         {
-            size_t match = whereView.get(i).getStr().find(charsView.get(i).getStr()[j], pos_start);
+            size_t match = whereView.get(i).getStr().find(utf8Char, pos_start);
 
             if (match <= pos_last)
                 current.emplace_back(match+1);
@@ -2304,7 +2359,7 @@ mu::Array strfnc_locate(const mu::Array& arr, const mu::Array& tofind, const mu:
             else if (t == 5)
             {
                 // Search any of the characters in the string
-                if (arg.find_first_of(sView.to_string()) != std::string::npos)
+                if (strmatch_impl(arg.to_string(), sView.to_string()) != std::string::npos)
                     ret.emplace_back(i+1);
             }
             else
@@ -2516,11 +2571,11 @@ mu::Array strfnc_findtoken(const mu::Array& sStr, const mu::Array& tok, const mu
     {
         StringView sView1 = strView.get(i).getStr();
         const std::string& t = tokView.get(i).getStr();
-        std::string s = " \t";
+        std::vector<std::string> separators({" ", "\t"});
 
         // Define default arguments
         if (!sep.isDefault())
-            s = sepView.get(i).getStr();
+            separators = splitUtf8Chars(sepView.get(i).getStr());
 
         size_t nMatch = 0;
 
@@ -2528,8 +2583,8 @@ mu::Array strfnc_findtoken(const mu::Array& sStr, const mu::Array& tok, const mu
         // defined separator characters
         while ((nMatch = sView1.find(t, nMatch)) != std::string::npos)
         {
-            if ((!nMatch || s.find(sView1[nMatch-1]) != std::string::npos)
-                && (nMatch + t.length() >= sView1.length() || s.find(sView1[nMatch+t.length()]) != std::string::npos))
+            if ((!nMatch || matchesAny(sView1, separators, findCharStart(sView1, nMatch-1)))
+                && (nMatch + t.length() >= sView1.length() || matchesAny(sView1, separators, nMatch+t.length())))
             {
                 ret.emplace_back(mu::Value(nMatch+1));
                 break;
@@ -2916,9 +2971,7 @@ mu::Array strfnc_justify(const mu::Array& arr, const mu::Array& align)
 
             // Remove surrounding whitespaces
             sStr.strip();
-
-            if (sStr.length() > maxLength)
-                maxLength = sStr.length();
+            maxLength = std::max(maxLength, countUnicodePoints(sStr));
         }
 
         // Fill all string with as many whitespaces as necessary
@@ -2928,15 +2981,16 @@ mu::Array strfnc_justify(const mu::Array& arr, const mu::Array& align)
             view.strip();
 
             std::string sStr = view.to_string();
+            size_t len = countUnicodePoints(sStr);
 
             if (a == 1)
-                sStr.insert(0, maxLength - sStr.size(), ' ');
+                sStr.insert(0, maxLength - len, ' ');
             else if (a == -1)
-                sStr.append(maxLength - sStr.size(), ' ');
+                sStr.append(maxLength - len, ' ');
             else if (a == 0)
             {
-                size_t leftSpace = (maxLength - sStr.size()) / 2;
-                size_t rightSpace = maxLength - leftSpace - sStr.size();
+                size_t leftSpace = (maxLength - len) / 2;
+                size_t rightSpace = maxLength - leftSpace - len;
                 sStr.insert(0, leftSpace, ' ');
                 sStr.append(rightSpace, ' ');
             }
@@ -3632,14 +3686,14 @@ mu::Array strfnc_valtostr(const mu::Array& vals, const mu::Array& cfill, const m
         if (!len.isDefault()
             && !cfill.isDefault()
             && fillView.get(i).getStr().length()
-            && (int64_t)v.length() < lenView.get(i).getNum().asI64())
+            && (int64_t)countUnicodePoints(v) < lenView.get(i).getNum().asI64())
         {
             int64_t l = lenView.get(i).getNum().asI64();
             const std::string& sChar = fillView.get(i).getStr();
 
-            while ((int64_t)v.length() < l)
+            while ((int64_t)countUnicodePoints(v) < l)
             {
-                v.insert(0, 1, sChar.front());
+                v.insert(0, sChar);
             }
         }
