+++ kernel/kernel.cpp
@@ -848,6 +848,8 @@ void NumeReKernel::defineStrFunctions()
     _parser.DefineFun("to_lowercase", strfnc_to_lowercase);                          // to_lowercase(str)
     _parser.DefineFun("to_ansi", strfnc_utf8ToAnsi);                                 // to_ansi(str)
     _parser.DefineFun("to_utf8", strfnc_ansiToUtf8);                                 // to_utf8(str)
+    _parser.DefineFun("to_codepoints", strfnc_to_codepoints);                        // to_codepoints(str)
+    _parser.DefineFun("from_codepoints", strfnc_from_codepoints);                    // from_codepoints({val})
     _parser.DefineFun("to_html", strfnc_toHtml);                                     // to_html(val)
     _parser.DefineFun("getenvvar", strfnc_getenvvar);                                // getenvvar(str)
     _parser.DefineFun("getfileparts", strfnc_getFileParts);                          // getfileparts(str)
@@ -891,6 +893,7 @@ void NumeReKernel::defineStrFunctions()
     _parser.DefineFun("timeformat", strfnc_timeformat);                              // timeformat(str,time)
     _parser.DefineFun("weekday", strfnc_weekday, true, 1);                           // weekday(d,opts)
     _parser.DefineFun("char", strfnc_char);                                          // char(str,n)
+    _parser.DefineFun("byte", strfnc_char);                                          // byte(str,n)
     _parser.DefineFun("getopt", strfnc_getopt);                                      // getopt(str,n)
     _parser.DefineFun("replace", strfnc_replace);                                    // replace(str,n,m,str)
     _parser.DefineFun("textparse", strfnc_textparse, true, 2);                       // textparse(str,str,p1,p2)
